Split getCatalog service fallback
Modify getCatalog of the movie-catalog-service in 2 separate calls to 2 separate services class.
The calls to services have to be in stand alone classes to allow to Hystrix to wrap a proxy around them. Putting methods with fallbacks in the same class of getCatalog wouldn't allow Hystrix to intercept service malfunctions or timeouts.